requirements.txt
allure-pytest==2.9.41
allure-python-commons==2.9.41
altgraph==0.17
apipkg==1.5
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
distlib==0.3.2
Eel==0.12.4
execnet==1.7.1
Faker==4.1.2
filelock==3.0.12
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
identify==2.2.10
idna==2.10
iniconfig==1.1.1
loguru==0.5.3
more-itertools==8.5.0
msedge-selenium-tools==3.141.3
nodeenv==1.6.0
packaging==20.9
pefile==2019.4.18
pluggy==0.13.1
pre-commit==2.13.0
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
pytest==6.2.4
pytest-forked==1.3.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.1.0
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.7
webdriver-manager==3.4.1
whichcraft==0.6.1
win32-setctime==1.0.3
zope.event==4.5.0
zope.interface==5.2.0
